requirements.txt
black==24.4.2
eth-account==0.13.0
eth-abi==5.2.0
eth-utils==4.1.1
poly_eip712_structs==0.0.1
py_order_utils==0.3.2
pytest==8.2.2
python-dotenv==0.19.2
py-builder-signing-sdk==0.0.2
httpx[http2]==0.27.0
